fix: address recording participants review feedback
Drop RecordingParticipants' own oatsRecordingByExternalId query -
RecordingDetailV2Screen already runs it, so recordingParticipants/shares
now flow down as props through the existing recordingRow merge effect.
Extract the component's participant/optimistic-state/search logic into
useRecordingParticipants so the component is just the JSX.

Services-Requiring-Redeployment:
  - dashboard
